refactor: add init to improve import ease & readability
